stress-bad-ioctl: cast -1 to size_t to clean up clang pedantic warning
clean up warning:
stress-bad-ioctl.c:536:38: warning: implicit conversion changes
signedness: 'int' to 'size_t' (aka 'unsigned long') [-Wsign-conversion]

+++ stress-bad-ioctl.c
@@ -533,7 +533,7 @@ static void stress_bad_ioctl_dir(
 				dev_ioctl_node = node;
 				(void)shim_memset(node->exercised, 0, sizeof(node->exercised));
 				(void)stress_lock_release(lock);
-				stress_bad_ioctl_rw(args, false, -1);
+				stress_bad_ioctl_rw(args, false, (size_t)-1);
 			}
 		}
 		stress_bogo_inc(args);
